# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PaikarMart multi-vendor eCommerce API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ProductType = (typeof ProductType)[keyof typeof ProductType];

export const ProductType = {
  physical: "physical",
  digital: "digital",
  service: "service",
} as const;

export interface Product {
  id: string;
  name: string;
  category: string;
  subcategory?: string;
  type: ProductType;
  price?: number;
  costPrice?: number;
  stock?: number;
  moq?: number;
  description?: string;
  images?: string[];
  vendorId: string;
  vendorName: string;
  location?: string;
  rating?: number;
  reviewCount?: number;
  priceOnInquiry?: boolean;
  inStock: boolean;
  deliveryDays?: string;
  createdAt?: string;
}

export interface ProductListResponse {
  products: Product[];
  total: number;
}

export interface Category {
  id: string;
  name: string;
  icon?: string;
  productCount?: number;
}

export interface CategoryListResponse {
  categories: Category[];
}

export type SellerBusinessType =
  (typeof SellerBusinessType)[keyof typeof SellerBusinessType];

export const SellerBusinessType = {
  wholesaler: "wholesaler",
  retailer: "retailer",
  brand_seller: "brand_seller",
  local_shop: "local_shop",
  dropship: "dropship",
  service: "service",
} as const;

export type SellerStatus = (typeof SellerStatus)[keyof typeof SellerStatus];

export const SellerStatus = {
  pending: "pending",
  approved: "approved",
  active: "active",
  suspended: "suspended",
} as const;

export interface Seller {
  id: string;
  shopName: string;
  businessType: SellerBusinessType;
  status: SellerStatus;
  phone?: string;
  email?: string;
  address?: string;
  location?: string;
  district?: string;
  rating?: number;
  totalProducts?: number;
  totalSales?: number;
  image?: string;
  description?: string;
  deliveryTypes?: string[];
  coverageAreas?: string[];
  createdAt?: string;
}

export interface SellerListResponse {
  sellers: Seller[];
  total: number;
}

export interface SellerRegistrationBody {
  shopName: string;
  businessType: string;
  phone: string;
  email?: string;
  address?: string;
  district?: string;
  categories?: string[];
}

export interface OrderItem {
  productId: string;
  productName: string;
  vendorId?: string;
  vendorName?: string;
  quantity: number;
  price: number;
  subtotal: number;
}

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

export const OrderStatus = {
  pending: "pending",
  confirmed: "confirmed",
  processing: "processing",
  shipped: "shipped",
  completed: "completed",
  incomplete: "incomplete",
} as const;

export type OrderDeliveryType =
  (typeof OrderDeliveryType)[keyof typeof OrderDeliveryType];

export const OrderDeliveryType = {
  seller_delivery: "seller_delivery",
  platform_delivery: "platform_delivery",
  local_delivery: "local_delivery",
  pickup: "pickup",
} as const;

export interface Order {
  id: string;
  userId?: string;
  items: OrderItem[];
  status: OrderStatus;
  deliveryType?: OrderDeliveryType;
  paymentMethod: string;
  totalAmount: number;
  deliveryCharge?: number;
  trackingCode?: string;
  customerName?: string;
  customerPhone?: string;
  customerAddress?: string;
  district?: string;
  area?: string;
  estimatedDelivery?: string;
  createdAt: string;
  updatedAt?: string;
}

export interface SellerDashboard {
  totalProducts: number;
  inStock: number;
  outOfStock: number;
  totalOrders: number;
  pendingOrders: number;
  completedOrders: number;
  totalSales: number;
  estimatedProfit: number;
  recentOrders: Order[];
}

export type CreateOrderBodyItemsItem = {
  productId: string;
  quantity: number;
};

export interface CreateOrderBody {
  items: CreateOrderBodyItemsItem[];
  customerName: string;
  customerPhone: string;
  customerEmail?: string;
  customerAddress: string;
  district?: string;
  area?: string;
  zipCode?: string;
  deliveryType?: string;
  paymentMethod: string;
}

export interface UpdateOrderStatusBody {
  status: string;
  trackingCode?: string;
  estimatedDelivery?: string;
}

export interface OrderListResponse {
  orders: Order[];
  total: number;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  buyer: "buyer",
  seller: "seller",
  admin: "admin",
} as const;

export interface User {
  id: string;
  name?: string;
  phone?: string;
  email?: string;
  role: UserRole;
  district?: string;
  area?: string;
  createdAt?: string;
}

export type WalletTransactionType =
  (typeof WalletTransactionType)[keyof typeof WalletTransactionType];

export const WalletTransactionType = {
  reward: "reward",
  adjustment: "adjustment",
} as const;

export interface WalletTransaction {
  id: string;
  type: WalletTransactionType;
  amount: number;
  description?: string;
  createdAt: string;
}

export interface Wallet {
  userId: string;
  balance: number;
  totalEarned: number;
  investmentValue: number;
  transactions: WalletTransaction[];
}

export type NotificationType =
  (typeof NotificationType)[keyof typeof NotificationType];

export const NotificationType = {
  order: "order",
  delivery: "delivery",
  account: "account",
  seller: "seller",
  system: "system",
} as const;

export interface Notification {
  id: string;
  userId?: string;
  type: NotificationType;
  title: string;
  message: string;
  read: boolean;
  createdAt: string;
}

export interface NotificationListResponse {
  notifications: Notification[];
  unreadCount: number;
}

export interface AdminDashboard {
  totalUsers: number;
  totalSellers: number;
  activeSellers: number;
  pendingSellers: number;
  totalOrders: number;
  totalSales: number;
  platformProfit: number;
  totalCost: number;
  recentOrders: Order[];
  recentSellers: Seller[];
}

export type MarketStatsMonthlyItem = {
  month: string;
  sales: number;
  cost: number;
  profit: number;
};

export type MarketStatsTrend =
  (typeof MarketStatsTrend)[keyof typeof MarketStatsTrend];

export const MarketStatsTrend = {
  up: "up",
  down: "down",
  stable: "stable",
} as const;

export interface MarketStats {
  monthly: MarketStatsMonthlyItem[];
  changePercent: number;
  trend: MarketStatsTrend;
}

export type ListProductsParams = {
  category?: string;
  vendor_id?: string;
  type?: string;
  search?: string;
  location?: string;
  min_price?: number;
  max_price?: number;
  limit?: number;
  offset?: number;
};

export type ListSellersParams = {
  type?: string;
  location?: string;
  status?: string;
};

export type ListOrdersParams = {
  user_id?: string;
  seller_id?: string;
  status?: string;
};

export type ListNotificationsParams = {
  user_id?: string;
  unread_only?: boolean;
};