# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: PaikarMart multi-vendor eCommerce API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders
      parameters:
        - name: user_id
          in: query
          schema:
            type: string
        - name: seller_id
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderListResponse"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Place an order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderBody"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Order detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}/status:
    put:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderStatusBody"
      responses:
        "200":
          description: Order updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
...
components:
  schemas:
...
    Order:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        status:
          type: string
          enum: [pending, confirmed, processing, shipped, completed, incomplete]
        deliveryType:
          type: string
          enum: [seller_delivery, platform_delivery, local_delivery, pickup]
        paymentMethod:
          type: string
        totalAmount:
          type: number
        deliveryCharge:
          type: number
        trackingCode:
          type: string
        customerName:
          type: string
        customerPhone:
          type: string
        customerAddress:
          type: string
        district:
          type: string
        area:
          type: string
        estimatedDelivery:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - items
        - status
        - totalAmount
        - paymentMethod
        - createdAt
    OrderItem:
      type: object
      properties:
        productId:
          type: string
        productName:
          type: string
        vendorId:
          type: string
        vendorName:
          type: string
        quantity:
          type: integer
        price:
          type: number
        subtotal:
          type: number
      required:
        - productId
        - productName
        - quantity
        - price
        - subtotal
    CreateOrderBody:
      type: object
      properties:
        items:
          type: array
          items:
            type: object
            properties:
              productId:
                type: string
              quantity:
                type: integer
            required:
              - productId
              - quantity
        customerName:
          type: string
        customerPhone:
          type: string
        customerEmail:
          type: string
        customerAddress:
          type: string
        district:
          type: string
        area:
          type: string
        zipCode:
          type: string
        deliveryType:
          type: string
        paymentMethod:
          type: string
      required:
        - items
        - customerName
        - customerPhone
        - customerAddress
        - paymentMethod
    UpdateOrderStatusBody:
      type: object
      properties:
        status:
          type: string
        trackingCode:
          type: string
        estimatedDelivery:
          type: string
      required:
        - status
    OrderListResponse:
      type: object
      properties:
        orders:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        total:
          type: integer
      required:
        - orders
        - total
...