# lib/api-zod/src/generated/types/createOrderBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PaikarMart multi-vendor eCommerce API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateOrderBodyItemsItem } from "./createOrderBodyItemsItem";

export interface CreateOrderBody {
  items: CreateOrderBodyItemsItem[];
  customerName: string;
  customerPhone: string;
  customerEmail?: string;
  customerAddress: string;
  district?: string;
  area?: string;
  zipCode?: string;
  deliveryType?: string;
  paymentMethod: string;
}