# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PaikarMart multi-vendor eCommerce API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderDeliveryType } from "./orderDeliveryType";
import type { OrderItem } from "./orderItem";
import type { OrderStatus } from "./orderStatus";

export interface Order {
  id: string;
  userId?: string;
  items: OrderItem[];
  status: OrderStatus;
  deliveryType?: OrderDeliveryType;
  paymentMethod: string;
  totalAmount: number;
  deliveryCharge?: number;
  trackingCode?: string;
  customerName?: string;
  customerPhone?: string;
  customerAddress?: string;
  district?: string;
  area?: string;
  estimatedDelivery?: string;
  createdAt: string;
  updatedAt?: string;
}