# lib/api-zod/src/generated/types/orderItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PaikarMart multi-vendor eCommerce API
 * OpenAPI spec version: 0.1.0
 */

export interface OrderItem {
  productId: string;
  productName: string;
  vendorId?: string;
  vendorName?: string;
  quantity: number;
  price: number;
  subtotal: number;
}